chore: separate messaging into separate file
